+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Game/Tagger.cpp
@@ -46,6 +46,7 @@ crp::CityObjectLabel ATagger::GetLabelByFolderName(const FString &String) {
   else if (String == "Dynamic")      return crp::CityObjectLabel::Dynamic;
   else if (String == "Water")        return crp::CityObjectLabel::Water;
   else if (String == "Terrain")      return crp::CityObjectLabel::Terrain;
+  else if (String == "Parking")      return crp::CityObjectLabel::Parking;
   else                               return crp::CityObjectLabel::None;
 }
 
@@ -251,6 +252,7 @@ FString ATagger::GetTagAsString(const crp::CityObjectLabel Label)
     CARLA_GET_LABEL_STR(Dynamic)
     CARLA_GET_LABEL_STR(Water)
     CARLA_GET_LABEL_STR(Terrain)
+    CARLA_GET_LABEL_STR(Parking)
 #undef CARLA_GET_LABEL_STR
   }
 }