+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * Fixed bug causing the TM to block the simulation when another client teleported a vehicle with no physics.
   * Added check to avoid adding procedural trigger boxes inside intersections.
+  * Added new tag for parking areas (Parking)
   * Python agents now accept a carla.Map and GlobalRoutePlanner instances as inputs, avoiding the need to recompute them.
   * Fix a bug at `Map.get_topology()`, causing lanes with no successors to not be part of it.
   * Added new ConstantVelocityAgent